pyproject.toml
[tool.poetry]
name = "pythoneda-infrastructure-base"
version = "0.0.1a12"
description = "Infrastructure layer for PythonEDA"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-infrastructure/base"
packages = [{include = "pythonedainfrastructure"}]

[tool.poetry.dependencies]
dbus-next = "^0.2.3"
python = "^3.8"
pythoneda-base = "^0.0.1a15"
requests = "^2.28.1"
grpcio = "^1.41.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate-grpc = "scripts.make_grpc:main"
